backend/requirements.txt
fastapi
uvicorn
langchain
google-generativeai
python-dotenv
pydantic
langchain-community
faiss-cpu
pypdf
langchain_google_genai
requests